Email: Move all types in dedicated file
