+++ client/my-sites/email/email-providers-in-depth-comparison/comparison-table/index.tsx
@@ -1,18 +1,5 @@
 import { FunctionComponent } from 'react';
-import type { TranslateResult } from 'i18n-calypso';
-
-export type EmailProviderFeatures = {
-	header: string;
-	tools: TranslateResult;
-	storage: TranslateResult;
-	importing: TranslateResult;
-	support: TranslateResult;
-	selectCallback: () => void;
-};
-
-type ComparisonTableProps = {
-	emailProviders: EmailProviderFeatures[];
-};
+import type { ComparisonTableProps } from 'calypso/my-sites/email/email-providers-in-depth-comparison/types';
 
 const ComparisonTable: FunctionComponent< ComparisonTableProps > = () => {
 	return null;
